pyproject.toml
[project]
name = "wechat-recorder"
version = "0.1.0"
description = "微信聊天记录读取器 - 基于 UI Automation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "uiautomation==2.0.19",
    "pyautogui==0.9.54",
    "pyperclip==1.8.2",
    "keyboard==0.13.5",
    "pandas==2.1.4",
    "openpyxl==3.1.2",
    "pywin32==306",
    "python-dotenv==1.0.0",
    "loguru==0.7.2",
    "numpy>=1.26.4",
    "pillow>=11.3.0",
    "opencv-python>=4.11.0.86",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
